Fix: Duplicated declare
